Changed frame descriptor depth to a float value, in order to handle 12 bits-camera.

+++ Camera/include/frame_desc.hh
@@ -18,7 +18,7 @@ namespace camera
     /*! Height of the frame. */
     unsigned short         height;
     /*! Byte depth during acquisition. */
-    unsigned short         depth;
+    float         depth;
     /*! Size of pixels in micrometers. */
     float                pixel_size;
     /*! Endianness of bytes. */
